refactor(vo): 패키지 분리
