+++ src/test/java/domain/account/AccountTest.java
@@ -1,9 +1,10 @@
-package vo;
+package domain.account;
 
 import domain.result.ResultProfitRatio;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import vo.BettingMoney;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -17,5 +18,4 @@ class AccountTest {
         account.applyProfit(ResultProfitRatio.match(ratio));
         assertThat(account.getProfit()).isEqualTo(expected);
     }
-
 }
