+++ src/test/java/domain/name/NamesTest.java
@@ -2,6 +2,7 @@ package domain.name;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import vo.Name;
 
 import java.util.List;
