+++ src/test/java/domain/participant/ParticipantTest.java
@@ -1,9 +1,9 @@
 package domain.participant;
 
-import domain.card.Card;
+import vo.Card;
 import domain.card.CardNumber;
 import domain.card.CardShape;
-import domain.name.Name;
+import vo.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import vo.BettingMoney;
