+++ src/test/java/vo/NameTest.java
@@ -1,4 +1,4 @@
-package domain.name;
+package vo;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
